BSP-145: Refactored models for transform/update/validate
