--- src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/bulkscan/ValidationStatus.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.model.bulkscan;
-
-public enum ValidationStatus {
-    SUCCESS,
-    WARNINGS,
-    ERRORS
-}
