+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/bulkscan/validation/ValidationStatus.java
@@ -1,0 +1,7 @@
+package uk.gov.hmcts.reform.finrem.caseorchestration.model.bulkscan.validation;
+
+public enum ValidationStatus {
+    SUCCESS,
+    WARNINGS,
+    ERRORS
+}
