+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/bulkscan/validation/in/OcrDataValidationRequest.java
@@ -1,8 +1,9 @@
-package uk.gov.hmcts.reform.finrem.caseorchestration.model.bulkscan;
+package uk.gov.hmcts.reform.finrem.caseorchestration.model.bulkscan.validation.in;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Getter;
+import uk.gov.hmcts.reform.finrem.caseorchestration.model.bulkscan.OcrDataField;
 
 import javax.validation.constraints.NotEmpty;
 import java.util.List;
